Re-organize javascripts
