+++ app/javascript/shared/components/UserStore.js
