+++ app/javascript/shared/components/UsernameInput.jsx
