+++ app/javascript/shared/components/__tests__/UserStore.test.js
