+++ app/views/dashboards/show.html.erb
@@ -69,4 +69,4 @@
 </main>
 
 <%= javascript_packs_with_chunks_tag "archivedPostFilters", defer: true %>
-<%= javascript_packs_with_chunks_tag "autocompleteOrgMember", defer: true %>
+<%= javascript_packs_with_chunks_tag "dashboards/autocompleteOrgMember", defer: true %>
